+++ src/test/ExSln2/c99/gen/hal_LedArray.c
@@ -24,7 +24,23 @@ hal_Led * hal_LedArray_maybe_get_led(hal_LedArray * self, uint8_t index)
 {
     if (index < self->_leds_length)
     {
-        return self->_leds[index];
+        return PRIVATE_hal_LedArray__get_led(self, index);
     }
     return NULL;
+}
+
+void hal_LedArray_increment_index(hal_LedArray * self)
+{
+    self->_index = hal_LedArray_inc_wrap(self->_index, self->_leds_length);
+}
+
+// just an excuse for a static method :)    
+uint8_t hal_LedArray_inc_wrap(uint8_t value, uint8_t max)
+{
+    value++;
+    if (value >= max)
+    {
+        value = 0;
+    }
+    return value;
 }