+++ src/test/ExSln2/c99/gen/hal_LedArray.h
@@ -11,6 +11,7 @@ struct hal_LedArray
 {
     hal_Led * * _leds;
     uint8_t _leds_length;
+    uint8_t _index;
 };
 
 
@@ -22,3 +23,8 @@ void hal_LedArray_ctor(hal_LedArray * self, hal_Led * * leds, uint8_t leds_lengt
 hal_Led * PRIVATE_hal_LedArray__get_led(hal_LedArray * self, uint8_t index);
 
 hal_Led * hal_LedArray_maybe_get_led(hal_LedArray * self, uint8_t index);
+
+void hal_LedArray_increment_index(hal_LedArray * self);
+
+// just an excuse for a static method :)    
+uint8_t hal_LedArray_inc_wrap(uint8_t value, uint8_t max);